+++ src/lib/components/select-contract/ContractListDetail.tsx
@@ -3,70 +3,32 @@ import { useWallet } from "@cosmos-kit/react";
 import { matchSorter } from "match-sorter";
 import { useMemo, useState } from "react";
 
+import { ContractsTable } from "../table";
 import { TextInput } from "lib/components/forms";
-import { Loading } from "lib/components/Loading";
 import { DisconnectedState, EmptyState, ZeroState } from "lib/components/state";
 import { TagSelection } from "lib/components/TagSelection";
 import { INSTANTIATED_LIST_NAME } from "lib/data";
-import type { ContractLocalInfo, ContractListInfo } from "lib/stores/contract";
-import type { ContractAddr, HumanAddr, LVPair, Option } from "lib/types";
+import { useAdminByContractAddresses } from "lib/services/contractService";
+import type { ContractListInfo } from "lib/stores/contract";
+import type { ContractAddr, ContractInfo } from "lib/types";
 import { formatSlugName } from "lib/utils";
 
-import { ContractListReadOnlyTable } from "./ContractListReadOnlyTable";
-import { ContractListTable } from "./ContractListTable";
-
-interface FilteredListDetailProps {
-  contracts: ContractLocalInfo[];
-  isReadOnly?: boolean;
-  contractRemovalInfo?: LVPair;
-  onContractSelect?: (addr: ContractAddr) => void;
-}
-
-const FilteredListDetail = ({
-  contracts,
-  isReadOnly,
-  contractRemovalInfo,
-  onContractSelect = () => {},
-}: FilteredListDetailProps) => {
-  if (contracts.length === 0)
-    return (
-      <EmptyState
-        image="https://assets.alleslabs.dev/illustration/search-not-found.svg"
-        message="No contracts match found. 
-        Make sure you are searching with contract address, name, or description."
-      />
-    );
-
-  return isReadOnly ? (
-    <ContractListReadOnlyTable
-      contracts={contracts}
-      onContractSelect={onContractSelect}
-    />
-  ) : (
-    <ContractListTable
-      contracts={contracts}
-      contractRemovalInfo={contractRemovalInfo}
-    />
-  );
-};
-
-interface ContractListTableProps {
-  address: Option<HumanAddr>;
+interface ContractListContentProps {
   contractListInfo: ContractListInfo;
-  isLoading?: boolean;
-  isReadOnly?: boolean;
-  filteredContracts: ContractLocalInfo[];
-  onContractSelect?: (addr: ContractAddr) => void;
+  filteredContracts: ContractInfo[];
+  onContractSelect: (addr: ContractAddr) => void;
+  isLoading: boolean;
+  isReadOnly: boolean;
 }
 
 const ContractListContent = ({
-  address,
   contractListInfo,
-  isReadOnly,
   filteredContracts,
-  onContractSelect,
   isLoading,
-}: ContractListTableProps) => {
+  onContractSelect,
+  isReadOnly,
+}: ContractListContentProps) => {
+  const { address } = useWallet();
   const isInstantiatedByMe =
     contractListInfo.slug === formatSlugName(INSTANTIATED_LIST_NAME);
 
@@ -75,91 +37,111 @@ const ContractListContent = ({
       <DisconnectedState text="to see contracts you've previously instantiated." />
     );
   }
-  if (isLoading) {
-    return <Loading />;
-  }
-  if (contractListInfo.contracts.length === 0) {
-    return (
-      <ZeroState
-        list={{
-          label: contractListInfo.name,
-          value: contractListInfo.slug,
-        }}
-        isReadOnly={isReadOnly}
-      />
-    );
-  }
+
   return (
-    <FilteredListDetail
+    <ContractsTable
       contracts={filteredContracts}
-      isReadOnly={isReadOnly}
-      contractRemovalInfo={
-        contractListInfo.isContractRemovable
-          ? { label: contractListInfo.name, value: contractListInfo.slug }
-          : undefined
+      isLoading={isLoading}
+      emptyState={
+        !contractListInfo.contracts.length ? (
+          <ZeroState
+            list={{
+              label: contractListInfo.name,
+              value: contractListInfo.slug,
+            }}
+            isReadOnly={isReadOnly}
+          />
+        ) : (
+          <EmptyState
+            imageVariant="not-found"
+            message="No contracts match found. 
+  Make sure you are searching with contract address, name, or description."
+          />
+        )
       }
-      onContractSelect={onContractSelect}
+      onRowSelect={onContractSelect}
+      isReadOnly={isReadOnly}
+      withCTA={
+        isReadOnly
+          ? undefined
+          : {
+              removingContractList: contractListInfo.isContractRemovable
+                ? { label: contractListInfo.name, value: contractListInfo.slug }
+                : undefined,
+            }
+      }
     />
   );
 };
 
 interface ContractListDetailProps {
   contractListInfo: ContractListInfo;
-  isLoading?: boolean;
+  isLoading: boolean;
   isReadOnly?: boolean;
-  onContractSelect?: (addr: ContractAddr) => void;
+  onContractSelect: (addr: ContractAddr) => void;
 }
 
 export const ContractListDetail = ({
   contractListInfo,
   isLoading,
-  isReadOnly,
   onContractSelect,
+  isReadOnly = false,
 }: ContractListDetailProps) => {
+  const { data: admins = {} } = useAdminByContractAddresses(
+    isReadOnly
+      ? []
+      : contractListInfo.contracts.map((contract) => contract.contractAddress)
+  );
+
   const [searchKeyword, setSearchKeyword] = useState("");
   const [tagFilter, setTagFilter] = useState<string[]>([]);
-  const { address } = useWallet();
+
   const filteredContracts = useMemo(
     () =>
       matchSorter(contractListInfo.contracts, searchKeyword, {
         keys: ["name", "description", "label", "contractAddress"],
         sorter: (sortedItem) => sortedItem,
-      }).filter((contract) =>
-        tagFilter.every((tag) => contract.tags?.includes(tag))
-      ),
-    [contractListInfo.contracts, searchKeyword, tagFilter]
+      })
+        .filter((contract) =>
+          tagFilter.every((tag) => contract.tags?.includes(tag))
+        )
+        .map<ContractInfo>((contractLocalInfo) => ({
+          ...contractLocalInfo,
+          admin: admins[contractLocalInfo.contractAddress],
+          latestUpdater: undefined,
+          latestUpdated: undefined,
+          remark: undefined,
+        })),
+    [admins, contractListInfo.contracts, searchKeyword, tagFilter]
   );
 
   return (
-    <Box minH="xs" pb={isReadOnly ? "0px" : "48px"}>
-      <Box px={isReadOnly ? "0px" : "48px"}>
-        <Flex gap={2} w="full" my={isReadOnly ? "24px" : "48px"}>
-          <TextInput
-            variant="floating"
-            value={searchKeyword}
-            setInputState={setSearchKeyword}
-            placeholder="Search with contract address, name, or description"
-            size={!isReadOnly ? "lg" : "md"}
+    <Box minH="xs">
+      <Flex gap={2} w="full" my={isReadOnly ? "24px" : "48px"}>
+        <TextInput
+          variant="floating"
+          value={searchKeyword}
+          setInputState={setSearchKeyword}
+          placeholder="Search with contract address, name, or description"
+          size={!isReadOnly ? "lg" : "md"}
+        />
+        {!isReadOnly && (
+          <TagSelection
+            result={tagFilter}
+            setResult={setTagFilter}
+            placeholder="No tag selected"
+            label="Filter by tag"
+            boxWidth="400px"
+            creatable={false}
           />
-          {!isReadOnly && (
-            <TagSelection
-              result={tagFilter}
-              setResult={setTagFilter}
-              placeholder="No tag selected"
-              label="Filter by tag"
-              boxWidth="400px"
-              creatable={false}
-            />
-          )}
-        </Flex>
-      </Box>
+        )}
+      </Flex>
       <ContractListContent
-        address={address as HumanAddr}
         contractListInfo={contractListInfo}
         filteredContracts={filteredContracts}
-        isReadOnly={isReadOnly}
-        onContractSelect={onContractSelect}
         isLoading={isLoading}
+        onContractSelect={onContractSelect}
+        isReadOnly={isReadOnly}
       />
     </Box>
   );