+++ src/lib/components/state/index.ts
@@ -1,4 +1,5 @@
 export * from "./DisconnectedState";
 export * from "./EmptyState";
 export * from "./InvalidState";
+export * from "./StateImage";
 export * from "./ZeroState";