+++ src/lib/components/table/contracts/ContractsTable.tsx
@@ -1,30 +1,42 @@
 import { TableContainer } from "../tableComponents";
 import { Loading } from "lib/components/Loading";
-import type { ContractInfo, Option } from "lib/types";
+import type { ContractAddr, ContractInfo, Option } from "lib/types";
 
 import { ContractsTableHeader } from "./ContractsTableHeader";
 import { ContractsTableRow } from "./ContractsTableRow";
+import type { CTAInfo } from "./ContractsTableRowCTA";
 
 interface ContractsTableProps {
   contracts: Option<ContractInfo[]>;
   isLoading: boolean;
   emptyState: JSX.Element;
+  onRowSelect: (contract: ContractAddr) => void;
+  isReadOnly?: boolean;
+  withCTA?: CTAInfo;
 }
 
 export const ContractsTable = ({
   contracts,
   isLoading,
   emptyState,
+  onRowSelect,
+  isReadOnly = false,
+  withCTA,
 }: ContractsTableProps) => {
   if (isLoading) return <Loading />;
   if (!contracts?.length) return emptyState;
 
-  const templateColumns =
-    "150px minmax(250px, 1fr) 200px 150px minmax(250px, 300px) 70px";
+  const templateColumns = isReadOnly
+    ? "minmax(160px, 300px) minmax(300px, 3fr) minmax(200px, 2fr) 1fr"
+    : "160px minmax(300px, 3fr) minmax(200px, 2fr) 150px 250px 80px";
 
   return (
     <TableContainer>
-      <ContractsTableHeader templateColumns={templateColumns} />
+      <ContractsTableHeader
+        templateColumns={templateColumns}
+        isReadOnly={isReadOnly}
+        withCTA={withCTA}
+      />
       {contracts.map((contractInfo) => (
         <ContractsTableRow
           key={
@@ -36,6 +48,9 @@ export const ContractsTable = ({
           }
           contractInfo={contractInfo}
           templateColumns={templateColumns}
+          onRowSelect={onRowSelect}
+          isReadOnly={isReadOnly}
+          withCTA={withCTA}
         />
       ))}
     </TableContainer>