+++ src/lib/components/table/contracts/ContractsTableRow.tsx
@@ -1,44 +1,45 @@
-import { Flex, Text, Grid, IconButton, Box, chakra } from "@chakra-ui/react";
+import { Flex, Text, Grid } from "@chakra-ui/react";
 
 import { TableRow } from "../tableComponents";
-import { useInternalNavigate, useGetAddressType } from "lib/app-provider";
+import { useGetAddressType } from "lib/app-provider";
 import { ExplorerLink } from "lib/components/ExplorerLink";
-import { CustomIcon } from "lib/components/icon";
-import {
-  AddToOtherListModal,
-  SaveContractDetailsModal,
-} from "lib/components/modal";
-import type { ContractInfo } from "lib/types";
+import type { ContractAddr, ContractInfo } from "lib/types";
 import { RemarkOperation } from "lib/types";
-import { dateFromNow, formatUTC } from "lib/utils";
 
 import { ContractNameCell } from "./ContractNameCell";
+import type { CTAInfo } from "./ContractsTableRowCTA";
+import { ContractsTableRowCTA } from "./ContractsTableRowCTA";
 import { TagsCell } from "./TagsCell";
-
-const StyledIconButton = chakra(IconButton, {
-  baseStyle: {
-    display: "flex",
-    alignItems: "center",
-    fontSize: "22px",
-    borderRadius: "36px",
-  },
-});
 
 interface ContractsTableRowProps {
   contractInfo: ContractInfo;
   templateColumns: string;
+  onRowSelect: (contract: ContractAddr) => void;
+  isReadOnly: boolean;
+  withCTA?: CTAInfo;
 }
 
 const InstantiatorRender = ({
-  contractInfo: { remark, latestUpdater },
-}: Pick<ContractsTableRowProps, "contractInfo">) => {
+  contractInfo: { instantiator, remark, latestUpdater },
+  isReadOnly,
+}: {
+  contractInfo: ContractInfo;
+  isReadOnly: boolean;
+}) => {
   const getAddressType = useGetAddressType();
 
   /**
    * @remarks handle the case where the data is too old and cannot be found
    */
   if (!latestUpdater)
-    return (
+    return instantiator ? (
+      <ExplorerLink
+        value={instantiator}
+        type="user_address"
+        showCopyOnHover
+        isReadOnly={isReadOnly}
+      />
+    ) : (
       <Text variant="body2" color="text.dark">
         N/A
       </Text>
@@ -63,6 +64,7 @@ const InstantiatorRender = ({
             value={latestUpdater}
             type={updaterType}
             showCopyOnHover
+            isReadOnly={isReadOnly}
           />
         </Flex>
       );
@@ -72,6 +74,7 @@ const InstantiatorRender = ({
           value={latestUpdater}
           type={updaterType}
           showCopyOnHover
+          isReadOnly={isReadOnly}
         />
       );
     default:
@@ -86,93 +89,44 @@ const InstantiatorRender = ({
 export const ContractsTableRow = ({
   contractInfo,
   templateColumns,
-}: ContractsTableRowProps) => {
-  const navigate = useInternalNavigate();
+  onRowSelect,
+  isReadOnly,
+  withCTA,
+}: ContractsTableRowProps) => (
+  <Grid
+    templateColumns={templateColumns}
+    onClick={() => onRowSelect(contractInfo.contractAddress)}
+    _hover={{ bg: "pebble.900" }}
+    transition="all .25s ease-in-out"
+    cursor="pointer"
+    minW="min-content"
+  >
+    <TableRow>
+      <ExplorerLink
+        value={contractInfo.contractAddress}
+        type="contract_address"
+        showCopyOnHover
+        isReadOnly={isReadOnly}
+      />
+    </TableRow>
 
-  return (
-    <Grid
-      templateColumns={templateColumns}
-      onClick={() =>
-        navigate({
-          pathname: `/contract/${contractInfo.contractAddress}`,
-        })
-      }
-      _hover={{ bg: "pebble.900" }}
-      transition="all .25s ease-in-out"
-      cursor="pointer"
-      minW="min-content"
-    >
-      <TableRow>
-        <ExplorerLink
-          value={contractInfo.contractAddress}
-          type="contract_address"
-          showCopyOnHover
-        />
-      </TableRow>
+    <TableRow>
+      <ContractNameCell
+        contractLocalInfo={contractInfo}
+        isReadOnly={isReadOnly}
+      />
+    </TableRow>
 
-      <TableRow>
-        <ContractNameCell contractLocalInfo={contractInfo} />
-      </TableRow>
+    <TableRow>
+      <TagsCell contractLocalInfo={contractInfo} isReadOnly={isReadOnly} />
+    </TableRow>
 
-      <TableRow>
-        <TagsCell contractLocalInfo={contractInfo} />
-      </TableRow>
+    <TableRow>
+      <InstantiatorRender contractInfo={contractInfo} isReadOnly={isReadOnly} />
+    </TableRow>
 
-      <TableRow>
-        <InstantiatorRender contractInfo={contractInfo} />
-      </TableRow>
-
-      <TableRow>
-        <Flex
-          direction="column"
-          gap={1}
-          onClick={(e) => e.stopPropagation()}
-          cursor="text"
-        >
-          {contractInfo.latestUpdated ? (
-            <>
-              <Text variant="body2">
-                {formatUTC(contractInfo.latestUpdated)}
-              </Text>
-              <Text variant="body2" color="text.dark">
-                {`(${dateFromNow(contractInfo.latestUpdated)})`}
-              </Text>
-            </>
-          ) : (
-            <Text variant="body2" color="text.dark">
-              N/A
-            </Text>
-          )}
-        </Flex>
-      </TableRow>
-
-      <TableRow>
-        <Box onClick={(e) => e.stopPropagation()}>
-          {contractInfo.lists ? (
-            <AddToOtherListModal
-              contractLocalInfo={contractInfo}
-              triggerElement={
-                <StyledIconButton
-                  icon={
-                    <CustomIcon name="bookmark-solid" color="violet.light" />
-                  }
-                  variant="ghost-gray"
-                />
-              }
-            />
-          ) : (
-            <SaveContractDetailsModal
-              contractLocalInfo={contractInfo}
-              triggerElement={
-                <StyledIconButton
-                  icon={<CustomIcon name="bookmark" />}
-                  variant="ghost-gray"
-                />
-              }
-            />
-          )}
-        </Box>
-      </TableRow>
-    </Grid>
-  );
-};
+    {!isReadOnly && (
+      <ContractsTableRowCTA contractInfo={contractInfo} withCTA={withCTA} />
+    )}
+  </Grid>
+);